Support customisation of default colours in unix
Add a field to LogBackend which specifies a ColorConfig
If using colour, ensure that we use the value for the corresponding
level in the backend's ColorConfig if specified, falling back to the
default colors.

Expose the colour constants found in log_nix.go so that applications can
reference them in specifying their own colour config.

Add a new ConvertColors func to log_nix as a convenient way of
converting a list of numbers representing ansi codes into the strings
needed, without having to wrap each one in colorSeq or colorSeqBold
individually.

Also expose ColorSeq and ColorSeqBold to enable individual levels to
be easily set bold or otherwise if that behaviour is desired, however.